Conditionally bind against storage API
Pretty coarse-grained tests to build the bindings against older versions of
libvirt that don't have the storage API.

The tests only check if the types introduced by the storage API are
present, and removes any code that depends on the storage API if they are
not. With that, invoking any storage related ruby methods will result in
method missing errors.