Implementation of Libvirt::open_auth
Based on virConnectOpenAuth.  Unfortunately it's fairly
complicated to use, as you need to follow a lot of rules:

Arguments:
1)  With no arguments, open_auth will attempt to open a read/write
connection to the default system connection, using the
virConnectAuthPtrDefault callback to collect credentials.
2)  The first argument, if specified, should be a string that
contains the libvirt URI to connect to.
3)  The second argument, if specified, should be an array that
specifies various parameters to pass to the auth structure.  This
array *must* be 3 elements in length.  The first element is another
array that specifies which credentials the callback is willing to
support.  The second element is the callback, but note that since
ruby interprets "my_auth" as a function call, instead of a function
point, a symbol should be passed instead (i.e. :my_auth).  The
third argument specifies any additional user data to be passed
along into the callback.  The function callback should take
two parameters; the first is an array of hashes that represent the
credential structures, and the second is the userdata passed in
earlier.  Each of the credential hashes contains 5 elements:

type - this is the type of credential to be looked at now
prompt - a suggested prompt to show to the user
challenge - additional challenge information
defresult - A default result to use, if credentials could not be obtained
            (if this is 'nil', it should not be used)
result - an element to store the result of collecting credentials.  This
         will be interpreted as a string

4)  The third argument, if specified, is a flags parameter controlling
whether to open the connection read/write or read-only.  If flags is
0, the connection is read/write; if flags is Libvirt::CONNECT_RO, the
connection is read-only.

Signed-off-by: Chris Lalancette <clalance@redhat.com>