Much better error reporting.
Previously we were creating error objects with only a
default string of "Failed to execute function <blah>",
and placing the actual libvirt error message in the
exception object in the @libvirt_message member.  However,
most people never look at @libvirt_message, so the
default error message is rather useless for figuring out
why things failed.

With this new patch, we instead put the detailed error
message in both the @libvirt_message and in the default
string.  While it does make the default string somewhat
verbose, it is also much more useful for debugging problems.

While I was at it, I also removed the useless "msg" parameter
from the create_error function.  Nobody was using it.

Signed-off-by: Chris Lalancette <clalance@redhat.com>