Fix up definition of const's.
Unfortunately I made a very stupid error in trying to conditionally
define some of the constants in the classes.  I thought that
all of the flags in libvirt were defined with #define; however,
it turns out that they are all defined as enums.  What that leads
to is that the #ifdef around all of the conditionally compiled
flags were always empty, meaning that they weren't being
defined *at all*, even when they existed.

To fix this, introduce a new "have_const" into extconf.rb that
goes looking for these particular constants anywhere in
libvirt/libvirt.h.  With this in place we can properly discover
whether those flags exist.

Signed-off-by: Chris Lalancette <clalance@redhat.com>