Canonicalize error handling.
With this patch in place, we only ever raise a
Libvirt::Error, but we put additional information
in the object (libvirt_code, libvirt_message, and
libvirt_function).  The usage of the various error
object was haphazard before, and this cleans it up.

The downside is that if anyone is relying on specific
errors to be thrown, their code will get broken after
this patch.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>