Make sure not to leak memory allocated with ALLOC_N.
Since many ruby functions can end up throwing exceptions,
any time we allocate memory using ALLOC_N we have to be
prepared to do an rb_protect to check for exceptions.
If we then see an exception, we have to free up the memory
we allocated and then raise the exception.

Signed-off-by: Chris Lalancette <clalance@redhat.com>