Update *_set_parameters()
This does 2 things:
1)  Adds a "has_flag" variable to set_parameters().  If
has_flag is true, then the API is assumed to accept a
flag parameter, and it is appropriately checked.  If it
is zero, the code ensures that there is no flag parameter.
All of the callers of set_parameters() are also updated.
2)  Fix up the documentation for the *_set_parameters
calls so that they properly reflect what arguments
they'll accept.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>