Deprecate some old names, and add replacements.
We don't really need the _get_ and _set_ on
the methods that we use; that is implied.  So
deprecate the methods that are things like
"max_migration_set_speed" and replace them with
things like "max_migration_speed=".  Also deprecated
the "get" in method names.  Also update the
documentation to mention that these types of names
are now deprecated.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>