Use a better set_typed_parameters call.
With this, we individually define which values in the
input hash are available.  This is a bit more restrictive
than before; in particular, unknown values in the hash
won't be ignored like they used to be.  On the other
hand, this will ensure correctness much earlier, and
will reduce the round-trips to the libvirt daemon.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>