Fix possible buffer not terminated.
As pointed out by coverity, strncpy() does *not* guarantee
that a string will be null-terminated.  In this case,
this is a particular problem since the name of the input
key is user-generated, and thus could be used to overflow
the buffer.  To fix this we make sure to NULL-terminate
the buffer, and then to only use up to LENGTH-1 bytes
in the strncpy.

Signed-off-by: Chris Lalancette <clalancette@irobot.com>