extconf: Don't try to detect macros with have_const()
have_const() can detect enum values correctly, but it won't work
for things like VIR_MIGRATE_PARAM_* which are defined as
preprocessor macros.

We could use have_macro() instead but, unlike the other have_*()
functions, that one doesn't add any symbol to extconf.h when the
macro is found, so it's useless for the purpose of conditionally
compiling some code.

Luckily, all of the macros we are trying to detect have been
part of libvirt for a very long time, so we can just assume that
they are present and still build correctly on all the platforms
we target.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>