extconf: Don't reset $DEFLIBPATH
The intention here was to make sure that a system-wide copy of
libvirt would not be picked up instead of the one that the command
line options point to; however, since the bindings don't only link
against libvirt itself but also against other libraries, by doing
so we can make it impossible to correctly link the module and,
even before we get to that point, to detect the availability of
functions.

For example, when building on macOS with

  $ rake build -- \
    --with-libvirt-include=$(brew --prefix libvirt)/include \
    --with-libvirt-lib=$(brew --prefix libvirt)/lib

detecting functions will fail with

  have_func: checking for virStorageVolWipe()
                      in libvirt/libvirt.h... ----- no

  clang [...]
        -I/usr/local/Cellar/ruby/3.0.0_1/include/ruby-3.0.0
        -I/usr/local/opt/libvirt/include -L/usr/local/opt/libvirt/lib
        -lruby.3.0 -lvirt
  ld: library not found for -lruby.3.0
  clang: error: linker command failed with exit code 1

Note how the include path for Ruby is present on the compiler
command line, but the library path for it isn't.

We can simply not reset $DEFLIBPATH and that will work correctly
anyway, because we add the paths provided on the command line to
$LIBPATH which takes precedence over $DEFLIBPATH.

https://gitlab.com/libvirt/libvirt-ruby/-/issues/3

Signed-off-by: Andrea Bolognani <abologna@redhat.com>