extconf: Move sanity check and make it blocking
Whether we've found libvirt automatically using pkg-config or
thanks to explicit include and library paths provided manually
by the user, it's useful to perform a quick sanity check to
verify that the libvirt we've found is actually usable.

If the library is not usable, fail early instead of proceeding
to check the availability of dozens of symbols.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>