domain.c: Use sensible default for libvirt_domain_qemu_agent_command
Usually, the default value of zero for optional arguments works
just fine. But that's not the case for
libvirt_domain_qemu_agent_command and its 'timeout' argument
because no timeout is usually not enough time for the guest agent
to reply (esp. when a more complex command was executed, e.g.
"guest-network-get-interfaces").

Use QEMU_AGENT_COMMAND_DEFAULT as the default value, which lets
libvirt use its own default timeout which is 5 seconds currently.

Closes: https://gitlab.com/libvirt/libvirt-ruby/-/issues/5
Signed-off-by: Michal Privoznik <mprivozn@redhat.com>