storage: Fix StoragePool#list_all_volumes
The vol_new() function, just like all other similar functions,
expects the VALUE passed as second argument to be a Connection.

Most callers get this right and use ruby_libvirt_conn_attr(p)
to extract the Connection VALUE from the StoragePool VALUE p,
so that it can be copied over when building the StorageVol.

In the case of StoragePool#list_all_volumes, however, the
StoragePool VALUE itself was passed to vol_new(). Subsequent
calls to any method on the returned StorageVols would result
in

  Expected Connection object (ArgumentError)

being raised.

Closes: https://gitlab.com/libvirt/libvirt-ruby/-/issues/6

Signed-off-by: Andrea Bolognani <abologna@redhat.com>