tests: Skip many tests by default
The test driver, which we're now using by default, only
implements a subset of the functionality of the QEMU driver,
so many test cases are now failing. Skip them.

Also skipped by default are a number of setup/teardown commands
that modify the host. We don't want that to happen unless the
user explicitly requested it.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>