tests: Introduce test_sleep()
The test suite is littered with calls to sleep(), which makes
running it quite slow. Those are arguably bad practice in any
scenario, but when running against an actual libvirt instance
they can serve as a useful crutch. If we're using the test
driver however, as is now the default, they're completely
pointless.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>