tests: Drop some commented out code
The FIXME states that virDomainDestroyFlags() is not supported
by the test driver, which was probably the case at the time when
it was written but is no longer true today.

Either way, the qemu-monitor-command API is unsurprisingly
exclusive to the QEMU driver, so it will never be possible to
run this code.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>