ext: Call rb_undef_alloc_func() for wrapper classes
Recent versions of Ruby (3.2 and newer) have started raising
warnings when running code that uses ruby-libvirt:

  undefining the allocator of T_DATA class Libvirt::Connect

This is caused by the Ruby change tracked here:

  https://bugs.ruby-lang.org/issues/18007

Basically, Ruby objects that wrap C structs (T_DATA) are treated
differently from pure Ruby objects (T_OBJECT) even though in
both cases the type passed to rb_define_class() is rb_cObject.

Specifically, the former don't need memory to be allocated for
them, and they are supposed to communicate this to the Ruby
runtime by undefining the allocator. If that doesn't happen, as
is currently the case for ruby-libvirt, then the allocator will
automatically be undefined and a warning will be raised.

In order to silence the warnings, call rb_undef_alloc_func()
for all classes to which it applies: namely, those that wrap
C structs, for example Domain, but not those that are native
Ruby objects, for example DomainInfo.

Closes: https://gitlab.com/libvirt/libvirt-ruby/-/issues/10

Signed-off-by: Andrea Bolognani <abologna@redhat.com>