rake: Clean up spec
Various improvements. Notably, some additional metadata is
added, which will result in a few useful links showing up on
the RubyGems.org page; the 'rubyforge_project' key is no longer
set, addressing a warning that was raised at build time;
authors are listed as separate entries in an array; and a
better short description is provided.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>