rake: Clean up file lists
There are two guiding principles behind these changes: avoid
globs, requiring files to be listed explicitly instead (this is
already largely the case), and avoid repeating the same path
verbatim more than once.

As a result of this clean up, two issues are fixed: the first
one is that header files are now listed as sources for the
library, which means that a build will happen if they are
changed; the other one is that extconf.h will no longer be
included in the release archives if 'rake build' just so
happens to have been executed before 'rake package'. This file
is generated as part of the build process, just like the
Makefile is, and doesn't belong in a release archive.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>