examples: Move to top-level directory
For someone who just cloned the repository or is browsing it
on GitLab, having the examples front and center rather than
buried in the doc/ directory makes things a lot more
discoverable.

From the website we can link to GitLab instead of showing the
raw source, which gets us niceties such as syntax highlighing.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>