site: Update and reduce in scope
The current site contains outdated information, as well as
information that can already be found elsewhere. The navigation
bar also looks different than the ones on the libvirt-php and
libvirt-java sites.

Adopt a policy of linking to relevant content wherever
possible, which reduces the scope significantly. This results
in a very nimble web presence which will hopefully not become
outdated anytime soon.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>