ext: Explicitly disallow use of 'new' for wrapper classes
Commit 27907495e7ecf8122b0ea3e66e3fa34d2714b5a8 has disabled
the allocator function for all wrapper classes in order to
shut down the warnings that Ruby had started emitting in recent
versions.

As a side effect of this change, it has become impossible to
use 'new' to create instances of said classes. This would be a
pretty serious regression, but in reality the naked constructor
never made sense in this context, since wrapper objects need a
corresponding C pointer to be of any use.

Even at the C level, for most libvirt classes there doesn't
exist a way to allocate an empty object, and instead objects
are returned by methods called on another object, typically
a Libvirt::Connect.

To make things better for users, raise a somewhat more
descriptive error message when 'new' is used against a wrapper
class.

Closes: https://gitlab.com/libvirt/libvirt-ruby/-/issues/12

Signed-off-by: Andrea Bolognani <abologna@redhat.com>